Remove legacy NumPy print options fixture and update doctests for NumPy 2.0
Remove the handle_np2 fixture from src/skimage/conftest.py that set
np.set_printoptions(legacy='1.21'). Since we now require numpy>=2.1,
the legacy mode is no longer needed.

Update all doctest expected values to reflect NumPy 2.0 scalar output:
- Scalars now display as np.float64(x), np.int64(x), np.True_, np.False_
- Tuples containing scalars show mixed types

Fixes:
- io/collection.py: Fixed variable name bug (image_col -> ic)

All 300 doctests pass with 7 skipped (for optional dependencies).

+++ src/_skimage2/_shared/utils.py
@@ -815,7 +815,7 @@ def safe_as_int(val, atol=1e-3):
     Examples
     --------
     >>> safe_as_int(7.0)
-    7
+    np.int64(7)
 
     >>> safe_as_int([9, 4, 2.9999999999])
     array([9, 4, 3])
@@ -826,7 +826,7 @@ def safe_as_int(val, atol=1e-3):
     ValueError: Integer argument required but received 53.1, check inputs.
 
     >>> safe_as_int(53.01, atol=0.01)
-    53
+    np.int64(53)
 
     """
     mod = np.asarray(val) % 1  # Extract mantissa
