+++ src/skimage/graph/_rag.py
@@ -37,7 +37,7 @@ def _edge_generator_from_csr(csr_array):
     >>> csr = sparse.csr_array(dense)
     >>> edges = _edge_generator_from_csr(csr)
     >>> list(edges)
-    [(0, 0, 1.0), (1, 1, 1.0)]
+    [(0, np.int32(0), np.float64(1.0)), (1, np.int32(1), np.float64(1.0))]
     """
     nrows = csr_array.shape[0]
     values = csr_array.data
